+++ CHANGELOG.md
@@ -37,6 +37,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### State Machine Breaking
+
+* [#211](https://github.com/babylonlabs-io/babylon/pull/211) Babylon inflation module
+
 ### Improvements
 
 * [#235](https://github.com/babylonlabs-io/babylon/pull/235) Change default values